% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bbox_to_sp.R
\name{bbox_to_sp}
\alias{bbox_to_sp}
\title{Creates and returns a bounding box out of a shape}
\usage{
bbox_to_sp(sp)
}
\arguments{
\item{sp}{A Spatial file}
}
\description{
Creates and returns a bounding box out of a shape
}
\author{
Jeffrey W Hollister
}

