% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calenderheat.R
\name{calendarHeat}
\alias{calendarHeat}
\title{Calendar Heatmap
# an R version of a graphic from:}
\source{
http://stat-computing.org/dataexpo/2009/posters/wicklin-allison.pdf
}
\usage{
calendarHeat(dates, values, ncolors = 99, color = "r2g",
  varname = "Values", date.form = "\%Y-\%m-\%d", ...)
}
\description{
Calendar Heatmap
# an R version of a graphic from:
}
\author{
Paul Bleicher
calendarHeat: An R function to display time-series data as a calendar heatmap
Copyright 2009 Humedica. All rights reserved.

This program is free software; you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation; either version 2 of the License, or
(at your option) any later version.

This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.

You can find a copy of the GNU General Public License, Version 2 at:
http://www.gnu.org/licenses/gpl-2.0.html
}

