% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/co.var.R
\name{co.var}
\alias{co.var}
\title{Calculates the coeffiecient of variation}
\usage{
co.var(...)
}
\arguments{
\item{x}{A vector}
}
\description{
Calculate the CV of a given vector
}
\author{
Martin Jung
}
\keyword{subset}

