% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/date2Julian.r
\name{date2Julian}
\alias{date2Julian}
\title{Function to convert dates to Julian days}
\usage{
date2Julian(dat)
}
\arguments{
\item{A}{date object
\code{FALSE}}
}
\description{
Function to convert dates to Julian days
}
\author{
Martin Jung
}

