% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/do_tukeyHowell.R
\name{do_tukeyHowell}
\alias{do_tukeyHowell}
\title{Do a Tukey or Games-Howell posthoc test}
\usage{
do_tukeyHowell(data, group, method = c("Tukey", "Games-Howell"))
}
\arguments{
\item{data}{A vector containing your data}

\item{method}{Which test ("Tukey" or "Games-Howell")}

\item{groups}{A Factor vector how the data is grouped}
}
\value{
returns Test results
}
\description{
Do a Tukey or Games-Howell posthoc test
}
\examples{
with(ChickWeight, {
 do_tukeyHowell(data = weight,
                group = Diet,
                method = "Tukey")
               }
               )
}
\author{
Anonymous
}

