% Generated by roxygen2 (4.0.2): do not edit by hand
\name{doSegments}
\alias{doSegments}
\title{Add segments to a standard scatterplot}
\usage{
doSegments(x, y, ll, eps = 0.05, ...)
}
\arguments{
\item{x}{x coordinate of scatter points}

\item{y}{y coordinates of scatter points}

\item{ll}{line length of segments. For instance SD or the standard error}

\item{eps}{hat width}
}
\description{
Uses the segments function to add segments to a point scatter plot.
}
\author{
Martin Jung
}
\keyword{distribution}

