% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Dynamite.R
\name{Dynamite}
\alias{Dynamite}
\title{One of those dynamite-ignition-device plots}
\usage{
Dynamite(height, error, names = NA, significance = NA, ylim = c(0,
  maxLim), ...)
}
\arguments{
\item{height}{Height of the barplot}

\item{error}{The length of the error line}

\item{names}{(Optional) vector containing names for x}

\item{significance}{(Optional) vector containingy significance stars for labeling}

\item{ylim}{(Optional) Length of y-axis}
}
\description{
Plain simple. Just a barplot with arrows and optionally significance on top ;)
}
\author{
Martin Jung
}
\keyword{distribution}

