% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/do_friedmann.R
\name{friedman.test.with.post.hoc}
\alias{friedman.test.with.post.hoc}
\title{Pairwise Pairwise Multiple Multiple Comparison with Unequal N‘s and/or Variances}
\source{
Journal Journal of Educational Statistics Educational Statistics,Vol.1, ,Vol.1, No. 2, 1976, pp. 113 . 2, 1976, pp. 113-125
}
\usage{
\method{friedman.test}{with.post.hoc}(formu, data, to.print.friedman = T,
  to.post.hoc.if.signif = T, to.plot.parallel = T, to.plot.boxplot = T,
  signif.P = 0.05, color.blocks.in.cor.plot = T, jitter.Y.in.cor.plot = F)
}
\arguments{
\item{formu}{Formula of the shape:     Y ~ X | block}

\item{data}{long data.frame with three columns: [[ Y (numeric), X (factor), block (factor) ]]}

\item{to.print.friedman}{Print the results (boolean)}

\item{to.post.hoc.if.signif}{Do test only if significant (boolean)}

\item{to.plot.boxplot}{Plot a boxplot (boolean)}

\item{signif.P}{Significane level (float)}

\item{color.blocks.in.cor.plot}{Show colors in correlation plot (boolean)}

\item{jitter.Y.in.cor.plot}{Do jitter on y value in correlation plot (boolean)}
}
\value{
returns Test results
}
\description{
Note: This function doesn't handle NA's! In case of NA in Y in one of the blocks, then that entire block should be removed.
}
\author{
Anonymous
}

