% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmerDiag.R
\name{glmerDiag}
\alias{glmerDiag}
\title{Diagnostics for an lme4 merMod Model}
\usage{
glmerDiag(mod)
}
\arguments{
\item{mod}{A mixed effects model object}
}
\value{
returns Graphs and stats
}
\description{
Diagnostics for an lme4 merMod Model
}
\author{
Martin Jung
}

