% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kappa.mer.R
\name{kappa.mer}
\alias{kappa.mer}
\title{Calculates the kappa value for a given merMod model
Condition matrix of a models parameters using Laplace approximations}
\usage{
vif.mer(modelfit)
}
\arguments{
\item{fit}{A merMod object}

\item{scale}{standardize Input (default: T)}

\item{center}{center to the mean (default: F)}

\item{add.intercept}{Add intercept to kappa matrix (default: T)}

\item{exact}{Exact kappa values}
}
\description{
Calculates the kappa value for a given merMod model
Condition matrix of a models parameters using Laplace approximations
}
\author{
Austin F. Frank (https://github.com/aufrank/R-hacks)
}
\keyword{lme4}
\keyword{vif,}

