% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list_functions}
\alias{list_functions}
\title{List all functions in a package}
\usage{
list_functions(package, all.names = FALSE, pattern)
}
\arguments{
\item{package}{Which package to list functions from}

\item{all.names}{Include hidden functions (starting with ".")? Defaults to
\code{FALSE}}

\item{pattern}{A regex pattern to limit the results}
}
\description{
List all functions in a package
}
\author{
Noam Ross
}

