% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multSE.R
\name{minsamp}
\alias{minsamp}
\title{Calculate minimum sample size for each group from a given output}
\source{
Anderson, MJ and J Santanta-Garcon. 2015. "Measures of precision for dissimilarity-based multivariate analysis of ecological communities." Ecology Letters 18(1): 66-73.
}
\usage{
minsamp(output, group = NA, ignore.min = NULL)
}
\arguments{
\item{output}{}

\item{group}{A factor for a group (Default = 1)}

\item{ignore.min}{Number samples to be removed before calculation (Default = NULL)}
}
\value{
returns Test results
}
\description{
Adapted and combined from Jon Lefcheck's method
}
\author{
Jon Lefcheck, Martin Jung
}
\keyword{internal}

