% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function}
\usage{
summarySE(data,Value)
}
\arguments{
\item{ggplot}{objects can be passed in ..., or to plotlist (as a list of ggplot objects)}

\item{cols:}{Number of columns in layout}

\item{layout:}{A matrix specifying the layout. If present, 'cols' is ignored.}
}
\description{
If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.
}
\author{
Anonymous
}
\keyword{ggplot,plot,legend}

