% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multSE.R
\name{multSE}
\alias{multSE}
\title{Sampling precision via multivariate dissimilarity-based standard error estimates}
\source{
Anderson, MJ and J Santanta-Garcon. 2015. "Measures of precision for dissimilarity-based multivariate analysis of ecological communities." Ecology Letters 18(1): 66-73.
}
\usage{
multSE(D, group = 1, nresamp = 1000, permanova = FALSE, mins = TRUE,
  ...)
}
\arguments{
\item{D}{a dist object from species-site matrix}

\item{group}{A factor for a group (Default = 1)}

\item{nresamp}{Number of resamples (Default = 1000)}

\item{permanova}{Conduct a permanova (Default = T)}

\item{mins}{Estimate the minimum number of samples required (Default = T)}
}
\value{
returns Test results
}
\description{
Adapted and combined from Jon Lefcheck's method
}
\author{
Jon Lefcheck, Martin Jung
}

