% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getGoogleScholarCit.R
\name{ncites_scholar}
\alias{ncites_scholar}
\title{Retrieve number of citations of a paper in Google Scholar}
\usage{
ncites_scholar(user, paper)
}
\arguments{
\item{user}{character vector. The user ID in Google Scholar Citations. Obtain from author's profile website in google scholar (http://scholar.google.com/citations?user=...)}

\item{paper}{character vector. The paper ID in Google Scholar Citations. Copy from publication list in author's profile website.}
}
\description{
This function retrieves the number of citations of a given paper in Google Scholar.
}
\examples{
ncites_scholar(user="B7vSqZsAAAAJ", paper="d1gkVwhDpl0C")
}
\author{
F. Rodriguez-Sanchez
}

