% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize a numeric vector}
\usage{
normalize(x)
}
\arguments{
\item{x}{Input numeric vector}
}
\value{
returns a normalized vector
}
\description{
Normalizes an input numeric vector to a proportional scale (0 - 1)
}
\examples{
raw <- vector(2,5,2,6,11,6,4,3)

 normalize(raw)
}
\author{
Martin Jung
}

