% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normDataWithin.R
\name{normDataWithin}
\alias{normDataWithin}
\title{Norms the data within specified groups in a data frame
Norms the data within specified groups in a data frame; it normalizes each
 subject (identified by idvar) so that they have the same mean, within each group
 specified by betweenvars.}
\usage{
normDataWithin(data,Value)
}
\arguments{
\item{data}{A data frame}

\item{idvar:}{the name of a column that identifies each subject (or matched subjects)}

\item{measurevar:}{the name of a column that contains the variable to be summarized}

\item{betweenvars:}{a vector containing names of columns that are between-subjects variables}

\item{na.rm:}{a boolean that indicates whether to ignore NA's}
}
\description{
Norms the data within specified groups in a data frame
Norms the data within specified groups in a data frame; it normalizes each
 subject (identified by idvar) so that they have the same mean, within each group
 specified by betweenvars.
}
\author{
R Cookbook (http://www.cookbook-r.com/Manipulating_data/Summarizing_data/)
}
\keyword{loading}
\keyword{package}

