% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/normhist.R
\name{normhist}
\alias{normhist}
\title{Display a histogram with different density curves}
\usage{
normhist(yourvector)
}
\arguments{
\item{values}{Input numerical vector}

\item{d1}{Enter the name of a distribtion you want to fit your data to}

\item{d2}{Enter the name of a second distribtion you want to fit your data to}
}
\description{
This functions takes use of the fitdistr(..) function in the MASS package
and is able to display density curves on a standard histogramm as well as
their cumulative distribution.
}
\author{
Martin Jung
}
\keyword{distribution}

