% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/numbers2words.R
\name{numbers2words}
\alias{numbers2words}
\title{Converts a numeric object to a string (in english)}
\usage{
numbers2words(x)
}
\arguments{
\item{A}{numeric object
\code{FALSE}}
}
\description{
Converts a numeric object to a string (in english)
}
\author{
John Fox (http://tolstoy.newcastle.edu.au/R/help/05/04/2715.html)
}

