% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pairs2}
\alias{pairs2}
\title{A nicer version of the famous pairs plot}
\usage{
pairs(num.mat, diag.panel=panel.hist, lower.panel=panel.smooth, upper.panel=panel.cor)
pairs(num.mat, lower.panel=panel.smooth, upper.panel=panel.cor, method="kendall")
}
\arguments{
\item{values}{Input data.frame or matrix}
}
\description{
Just a wrapper for a better pairs plot using custom panels
Allows additional parameters for the panels, that can be found in
panel.hist , panel.smooth and panel.cor
}
\author{
Anonymous
}
\keyword{distribution,}
\keyword{pairs}

