% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/r2_nakagawa_lme4.R
\name{r2.nakagawa.lme4}
\alias{r2.nakagawa.lme4}
\title{Calculates Nakagawa conditional and marginal coefficient of determination for a lme4 model
Marginal r² gives the variance explained by fixed factors
Conditional r² gives the variance explained by fixed and random factors (entire model)}
\usage{
r2.nakagawa.lme4(m, varType = 2)
}
\arguments{
\item{A}{lme4 model object}

\item{Variance}{Type (default 2)
\code{FALSE}}
}
\description{
Calculates Nakagawa conditional and marginal coefficient of determination for a lme4 model
Marginal r² gives the variance explained by fixed factors
Conditional r² gives the variance explained by fixed and random factors (entire model)
}
\author{
Martin Jung
}

