% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remNa.R
\name{remNa}
\alias{remNa}
\title{Removes all rows with NA out of input}
\usage{
remNa(yourDatafram)
}
\arguments{
\item{x}{Input R object}
}
\description{
Makes a subset of the input
}
\author{
Martin Jung
}
\keyword{subset}

