% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remove_nodes.R
\name{rm_polygonnodes}
\alias{rm_polygonnodes}
\title{Simplifies a polygon by removing every n vertices}
\usage{
rm_polygonnodes(sp_obj, rep)
}
\arguments{
\item{sp_obj}{A spatial object}

\item{rep}{Which n nodes should be removed}
}
\description{
Simplifies a polygon by removing every n vertices
}
\author{
Jeffrey W Hollister
}

