% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setdiff2}
\alias{setdiff2}
\title{Show difference between two vectors}
\usage{
setdiff2(x,y)
}
\arguments{
\item{x}{vector x}

\item{y}{vector y}
}
\description{
returns a list of the elements of x that are not in y
and the elements of y that are not in x (not the same thing...)
}
\author{
Martin Jung
}
\keyword{difference}

