% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sp_RectangularBuffer.R
\name{sp_recbuf}
\alias{sp_recbuf}
\title{Creates a rectangular buffer around each point}
\usage{
sp_recbuf(points, radius = 20)
}
\arguments{
\item{points}{A spatialPoints}

\item{radius}{The radius according to the projection (Default: 20)}
}
\value{
Returns a rectangular polygon around each point
}
\description{
Creates a rectangular buffer around each point
}
\author{
Martin Jung
}

