% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/standardPackages.R
\name{standardPackages}
\alias{standardPackages}
\title{Load and/or installs any number of standardpackages}
\usage{
standardPackages(...)
}
\arguments{
\item{which}{(Optional) Stuff to load. Available: 'Standard' (Default), 'GIS', 'Plot' (Default), 'Analysis', 'Diversity', 'ALL'}
}
\description{
Just a specific function for me that quietly
loads all the packages I regularly use. Able to make subsets for specific uses
}
\author{
Martin Jung
}
\keyword{loading}
\keyword{package}

