% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stderr}
\alias{stderr}
\title{Calculates the Standard error of the mean}
\usage{
stderr(x)
}
\arguments{
\item{values}{Input numerical vector}
}
\description{
Calculates the standarderror of the mean
}
\author{
Martin Jung
}
\keyword{standard-error}

