% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summarySE.R
\name{summarySE}
\alias{summarySE}
\title{Summarizes data
Gives count, mean, standard deviation, standard error of the mean, and confidence interval (default 95%).}
\usage{
summarySE(data,Value)
}
\arguments{
\item{data}{A data frame}

\item{measurevar:}{the name of a column that contains the variable to be summarized}

\item{groupvars:}{a vector containing names of columns that contain grouping variables}

\item{na.rm:}{a boolean that indicates whether to ignore NA's}

\item{conf.interval:}{the percent range of the confidence interval (default is 95%)}
}
\description{
Summarizes data
Gives count, mean, standard deviation, standard error of the mean, and confidence interval (default 95%).
}
\author{
R Cookbook (http://www.cookbook-r.com/Manipulating_data/Summarizing_data/)
}
\keyword{summary}

