% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testdist.R
\name{testdist}
\alias{testdist}
\title{Attemptes to fit input values to different distributions.}
\usage{
testdist(yourvector)
}
\arguments{
\item{values}{Input numerical vector}
}
\description{
This functions takes use of the fitdistr(..) function in the MASS package
Returns a sorted table containing AIC/BIC values. Might give you some indication
of the underlying value distribution, but please do a visual expectation as well!
}
\author{
Martin Jung
}
\keyword{distribution}

