% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testRasterWriteSpeed.R
\name{testRasterWriteSpeed}
\alias{testRasterWriteSpeed}
\title{Print status messages tracking run time for an R script}
\usage{
testRasterWriteSpeed(rastsize = 8000, fileext = ".envi", tmpdir = ".",
  datatype = "FLT8S", reps = 10)
}
\arguments{
\item{rastsize}{the dimension of a side of a square matrix to write to disk
as a \code{RasterLayer}}

\item{tmpdir}{the directory to write the \code{RasterLayer} in}

\item{datatype}{the data type to use when writing the \code{RasterLayer}.
See \code{\link{raster::dataType}}}

\item{reps}{}

\item{filext}{the file extension to use for the temporary filename passed to
the \code{writeRaster} function}
}
\description{
Print status messages tracking run time for an R script
}
\examples{
#TODO: Add examples
}
\author{
Alex Zvoleff
}

