% Generated by roxygen2 (4.0.2): do not edit by hand
\name{TheMatrix}
\alias{TheMatrix}
\title{Interface to the Matrix in R}
\usage{
TheMatrix()
}
\arguments{
\item{x}{How many x points}

\item{y}{How many y numbers}
}
\description{
Possible use? Non. But 100% awesomeness ;)
}
\author{
Anonymous
}
\keyword{1337}

