% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/theme_minimal2.R
\name{theme_minimal2}
\alias{theme_minimal2}
\title{Custom minimal them for GGplot}
\usage{
theme_minimal2
}
\description{
Custom minimal them for GGplot
}
\author{
Anonymous
}
\keyword{ggplot,plot}

