% Generated by roxygen2 (4.0.2): do not edit by hand
\name{trim_whitespace}
\alias{trim_whitespace}
\title{Trims the whitespace from a string}
\usage{
trim_whitespace(s)
}
\arguments{
\item{s}{Input string}
}
\description{
Removes all whitespace from a string
}
\author{
Martin Jung
}
\keyword{manipulation}
\keyword{string}

