% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tryN.R
\name{tryN}
\alias{tryN}
\title{Wrapper for the default tryCatch call}
\usage{
tryN(f)
}
\arguments{
\item{A}{function
\code{FALSE}}
}
\description{
Wrapper for the default tryCatch call
}
\author{
Martin Jung
}

