% Generated by roxygen2 (4.0.2): do not edit by hand
\name{g_legend}
\alias{g_legend}
\title{Extracts the legend of a ggplot objects as grob}
\usage{
summarySE(data,Value)
}
\arguments{
\item{a}{ggplot object}
}
\description{
Extracts the legend of a ggplot objects as grob
}
\author{
Anonymous
}
\keyword{ggplot,plot,legend}

