% Generated by roxygen2 (4.0.2): do not edit by hand
\name{numbers_from_string}
\alias{numbers_from_string}
\title{Extract numbers from string or character}
\usage{
numbers_from_string("Today we ate 15 apples")
}
\arguments{
\item{x}{Input string}
}
\description{
Good for data mining
}
\author{
Martin Jung
}
\keyword{subset}

