% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/random_y_from_x.R
\name{random_y_from_x}
\alias{random_y_from_x}
\title{Samples a number of values from y}
\usage{
random_y_from_x(trees$Volume,10)
}
\arguments{
\item{x}{Input R vector}

\item{y}{Number of samples}
}
\description{
Makes a subset of the input
}
\author{
Martin Jung
}
\keyword{subset}

