% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/save.xlsx.R
\name{save.xlsx}
\alias{save.xlsx}
\title{Save a list of data.frames as new sheet in a new excel file}
\usage{
save.xlsx("myworkbook.xlsx", dataframe1, dataframe2, ...)
}
\arguments{
\item{file}{output file name}
}
\description{
This functions require the xlsx package to be installed and working.
You can save any number of dataframes into a new excel sheet
}
\author{
Martin Jung
}
\keyword{excel}
\keyword{export}

