% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stderr.R
\name{sem}
\alias{sem}
\title{Calculates the Standard error of the mean}
\usage{
sem(x)
}
\arguments{
\item{values}{Input numerical vector}
}
\description{
Calculates the standarderror of the mean
}
\author{
Martin Jung
}
\keyword{standard-error}

