% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vif.mer}
\alias{vif.mer}
\title{Calculates the variance inflation factor (VIF) of a model}
\usage{
vif.mer(modelfit)
}
\arguments{
\item{fit}{A merMod object}
}
\description{
Calculates the variance inflation factor (VIF) of a model
}
\author{
Austin F. Frank (https://github.com/aufrank/R-hacks)
}
\keyword{lme4}
\keyword{vif,}

